zb_backend/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
boto3==1.21.37
botocore==1.24.37
cycler==0.11.0
distlib==0.3.4
Django==4.0.3
djangorestframework==3.13.1
filelock==3.6.0
fonttools==4.31.2
importlib-metadata==4.11.3
jmespath==1.0.0
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
zipp==3.8.0
